lib: sbi_ecall: Probe based on device availability
Some SBI extensions depend on certain devices to be available. Without
these devices, the SBI calls do not work and should not report as
available when probed. There's already a probe function for
sbi_ecall_srst, so we add it for sbi_ecall_{ipi,rfence,time}.

An example where this is relevant is T-Head C906, which has an
incompatible CLINT timer, but nevertheless can support S-mode software
with its (Sstc-incompatible) stimecmp register. We do not want to
confuse S-mode software by reporting that we support SBI TIME.

Reported-by: Icenowy Zheng <uwu@icenowy.me>
Signed-off-by: Vivian Wang <dramforever@live.com>