firmware: Add FW_PIC_RELR option
RELR is a new packed format for R_*_RELATIVE relocations. It is
supported in LLD, but, as of this commit, BFD ld does not yet support it
for RISC-V.

As it is a relatively new feature, it is opt-in with a make option
FW_PIC_RELR=y. BFD ld supports only -z pack-relative-relocs but not yet
on RISC-V. Earlier versions of LLD supported this option with
--pack-dyn-relocs=relr, while later version also added support for -z
pack-relative-relocs. The Makefile tries to detect which option to use,
preferring -z pack-relative-relocs, assuming that BFD ld will add
support for it on RISC-V in the future.

A quick experiment shows that a default RV64 LLVM=1 build saves around
10 KiB of read only data by using FW_PIC_RELR=y.

Signed-off-by: Vivian Wang <dramforever@live.com>