firmware: Remove useless ALIGN
The .rela.dyn section should be exactly the size of the relocations,
without padding. On RV64, .rela* sections are already aligned and
there's no need for padding. On RV32, forcing __rel_dyn_end to be
aligned to 8 bytes confuses the relocation loop into processing an extra
entry past the end of .rela*, causing an invalid memory access.

Signed-off-by: Vivian Wang <dramforever@live.com>