firmware: Add support for Linux kernel image header
The firmwares fw_payload and fw_jump support the same entrypoint ABI as
M-mode Linux. To support booting theses firmware images from bootloaders
already supporting M-mode Linux, add a compatible header to the firmware
image.

The header starts with a jump to skip to the actual code, so including
it does not affect operation without a bootloader.

Signed-off-by: Vivian Wang <dramforever@live.com>